poc/example-6/requirements.txt
fastapi
psycopg2-binary
kubernetes